:wrench: [refactor] (#88) ChatActivity, ChatFragment에 Hilt/코루틴 아키텍처 적용
