+++ app/src/main/java/com/example/mediforme/ui/home/chat/ChatActivity.kt
@@ -6,35 +6,40 @@ import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.mediforme.remote.api.getRetrofit
 import com.example.mediforme.databinding.FragmentChatBinding
-import com.example.mediforme.remote.api.ChatGptApiService
-import com.example.mediforme.remote.api.ChatGptResponseDto
-import com.example.mediforme.remote.api.QuestionRequestDto
-import retrofit2.Call
-import retrofit2.Callback
+import com.example.mediforme.remote.api.ApiService
+import com.example.mediforme.remote.model.request.QuestionRequestDto
+import com.example.mediforme.remote.model.response.ApiResponse
+import com.example.mediforme.remote.model.response.ChatGptResponseDto
+import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.launch
 import retrofit2.Response
-import retrofit2.Retrofit
+import javax.inject.Inject
 
+// Hilt를 사용하여 의존성 주입을 활성화
+@AndroidEntryPoint
 class ChatActivity : AppCompatActivity() {
 
     private lateinit var binding: FragmentChatBinding
     private lateinit var chatAdapter: ChatAdapter
     private val messages = mutableListOf<Message>()
-    private lateinit var apiService: ChatGptApiService
     private lateinit var sharedPreferences: SharedPreferences
     private var userName:  String? = null // userName을 String 타입으로 선언
 
+    // Hilt가 ApiService 인스턴스를 자동으로 주입
+    @Inject
+    lateinit var apiService: ApiService
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = FragmentChatBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        // Retrofit 초기화 및 API 서비스 생성
-        val retrofit: Retrofit = getRetrofit()
-        apiService = retrofit.create(ChatGptApiService::class.java)
+//        // 코루틴과 의존성 주입방식을 사용 -> getRetrofit()을 호출하는 초기화 과정 생략
+//        val retrofit = getRetrofit()
+//        apiService = retrofit.create(ApiService::class.java)
 
         // RecyclerView 설정
         chatAdapter = ChatAdapter(messages)
@@ -87,25 +92,30 @@ class ChatActivity : AppCompatActivity() {
             return
         }
 
-        // API 호출
-        val questionRequest = QuestionRequestDto(question)
-        apiService.askQuestion(authToken, questionRequest).enqueue(object : Callback<ChatGptResponseDto> {
-            override fun onResponse(call: Call<ChatGptResponseDto>, response: Response<ChatGptResponseDto>) {
+        // 코루틴으로 API 호출
+        // lifecycleScope를 사용해 Activity의 생명주기에 맞춰 코루틴을 관리
+        lifecycleScope.launch {
+            try {
+                val questionRequest = QuestionRequestDto(question)
+                // suspend 함수를 호출하여 비동기 작업을 처리합니다.
+                val response: Response<ApiResponse<ChatGptResponseDto>> = apiService.askQuestion(authToken, questionRequest)
+
                 if (response.isSuccessful) {
-                    response.body()?.let {
-                        val answer = it.choices.firstOrNull()?.message?.content ?: "응답을 받지 못했습니다."
+                    val apiResponse = response.body()
+                    if (apiResponse != null && apiResponse.isSuccess) {
+                        val answer = apiResponse.result.choices.firstOrNull()?.message?.content ?: "응답을 받지 못했습니다."
                         receiveMessage(answer)
+                    } else {
+                        receiveMessage("API 응답 실패: ${apiResponse?.message ?: "알 수 없는 오류"}")
                     }
                 } else {
                     val errorBody = response.errorBody()?.string()
                     receiveMessage("서버 응답 실패: ${response.message()} \n오류 내용: $errorBody")
                 }
+            } catch (e: Exception) {
+                receiveMessage("오류 발생: ${e.message}")
             }
-
-            override fun onFailure(call: Call<ChatGptResponseDto>, t: Throwable) {
-                receiveMessage("오류 발생: ${t.message}")
-            }
-        })
+        }
     }
 
     private fun receiveMessage(message: String) {
