+++ app/src/main/java/com/example/mediforme/ui/home/chat/ChatFragment.kt
@@ -9,28 +9,34 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.mediforme.remote.api.getRetrofit
 import com.example.mediforme.databinding.FragmentChatBinding
-import com.example.mediforme.remote.api.ChatGptApiService
-import com.example.mediforme.remote.api.ChatGptResponseDto
-import com.example.mediforme.remote.api.QuestionRequestDto
+import com.example.mediforme.remote.api.ApiService
+import com.example.mediforme.remote.model.request.QuestionRequestDto
+import com.example.mediforme.remote.model.response.ApiResponse
+import com.example.mediforme.remote.model.response.ChatGptResponseDto
+import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.launch
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.Retrofit
+import javax.inject.Inject
 
+// Hilt를 사용하여 의존성 주입을 활성화
+@AndroidEntryPoint
 class ChatFragment : Fragment() {
 
     private lateinit var binding: FragmentChatBinding
     private lateinit var chatAdapter: ChatAdapter
     private val messages = mutableListOf<Message>()
-    private lateinit var apiService: ChatGptApiService
     private val authToken = "Bearer API키!!!!!!!!" // 여기에 실제 API 키를 입력하세요
     private lateinit var sharedPreferences: SharedPreferences
     private var userName:  String? = null // userName을 String 타입으로 선언
 
-
+    // 의존성 주입을 통해 ApiService 인스턴스를 받음
+    @Inject
+    lateinit var apiService: ApiService
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -46,10 +52,6 @@ class ChatFragment : Fragment() {
         Log.d("dddkkk","${userName}")
         // 사용자 이름을 chat_name_TV에 설정
         binding.chatNameTV.text = userName
-
-        // Retrofit 초기화 및 API 서비스 생성
-        val retrofit: Retrofit = getRetrofit()
-        apiService = retrofit.create(ChatGptApiService::class.java)
 
         // RecyclerView 설정
         chatAdapter = ChatAdapter(messages)
@@ -85,25 +87,28 @@ class ChatFragment : Fragment() {
         // 메시지 전송 후 EditText의 내용을 지움
         binding.sendET.text.clear()
 
-        // API 호출
-        val questionRequest = QuestionRequestDto(question)
-        apiService.askQuestion(authToken, questionRequest).enqueue(object : Callback<ChatGptResponseDto> {
-            override fun onResponse(call: Call<ChatGptResponseDto>, response: Response<ChatGptResponseDto>) {
+        // 코루틴으로 API 호출
+        lifecycleScope.launch {
+            try {
+                val questionRequest = QuestionRequestDto(question)
+                val response: Response<ApiResponse<ChatGptResponseDto>> = apiService.askQuestion(authToken, questionRequest)
+
                 if (response.isSuccessful) {
-                    response.body()?.let {
-                        val answer = it.choices.firstOrNull()?.message?.content ?: "응답을 받지 못했습니다."
+                    val apiResponse = response.body()
+                    if (apiResponse != null && apiResponse.isSuccess) {
+                        val answer = apiResponse.result.choices.firstOrNull()?.message?.content ?: "응답을 받지 못했습니다."
                         receiveMessage(answer)
+                    } else {
+                        receiveMessage("API 응답 실패: ${apiResponse?.message ?: "알 수 없는 오류"}")
                     }
                 } else {
                     val errorBody = response.errorBody()?.string()
                     receiveMessage("서버 응답 실패: ${response.message()} \n오류 내용: $errorBody")
                 }
+            } catch (e: Exception) {
+                receiveMessage("오류 발생: ${e.message}")
             }
-
-            override fun onFailure(call: Call<ChatGptResponseDto>, t: Throwable) {
-                receiveMessage("오류 발생: ${t.message}")
-            }
-        })
+        }
     }
 
     private fun receiveMessage(message: String) {
